test: cobre falha de leitura do perfil
